Refactor: Footer 디렉터리 안에 FooterTopContainer 디렉터리를 추가 - Footer 컴포넌트의 파일 구조를 한눈에 알아볼 수 있도록 Container로 나눠지는 부분에 디렉터리를 추가
